+++ modules/belous.tetris.game/belous/tetris/game/api/Player.java
@@ -2,5 +2,5 @@ package belous.tetris.game.api;
 
 public interface Player {
     Move makeMove(State state);
-    void stateUpdated(State state, long score);
+    void stateUpdated(State state, int score);
 }
